regression to mean stuff
looking into the wide_clickbot attitude_change data to see if regression to the mean might explain our changes. I don't think so? Because the againsts increased less than the neutrals, and stayed the same more than the neutrals, with a similar number decreasing , so the change is driven by the neutrals not the againsts? so not due to us sampling the lower end and them regressing to the middle?